NMS-10533 move to DAOs instead of plain SQL
