+++ opennms-dao-api/src/main/java/org/opennms/netmgt/dao/api/OnmsDao.java
@@ -57,6 +57,11 @@ public interface OnmsDao<T, K extends Serializable> {
 
     void delete(K key);
 
+    /**
+     * Deletes all entities of this type - use with care!
+     */
+    void deleteAll();
+
     List<T> findAll();
     
     List<T> findMatching(Criteria criteria);
